Apply custom patch for BASWSPRT-92
The patch was originaly done here :

https://github.com/colemanw/webform_civicrm/pull/280

But due to the diffrence in the way we handle the first contribution
amount currently, it will not be merged.

The changes I did here are minimal unlike the PR above, I only make sure
that the recur contribution amount will always equal the first
contribution amount.