ASESPRT-390: Fix Sagepay payment processor payments
In CiviCRM  versions > 5.28.3 , the way in which something called qf key is changed where it now contains a prefix:

https://github.com/civicrm/civicrm-core/blob/5.35.1/CRM/Core/Key.php#L159-L160

The thing is when we submit a Sagepay payment through webforms, it will result in calling the following method
https://github.com/compucorp/webform_civicrm/blob/7.x-4.28-patches/includes/wf_crm_webform_postprocess.inc#L1937 and one of the things this method does is that it generates
parameters that are going to be used when generating the notification callback URL for sagepay, and one of these parameters is the qf key:
https://github.com/compucorp/webform_civicrm/blob/7.x-4.28-patches/includes/wf_crm_webform_postprocess.inc#L1957

In CiviCRM versions <= 5.28.3, the notification URL would end up to look something like this :

https://test-site-url.co.uk/civicrm/payment/ipn?processor_id=5&cid=62817&conid=36637&mo=contribute&v=testprov&qf=cdc54b72d5f869d758ba719f5474e8e3ec280fa68b3d7deaf520797d4f52cb4c_9839

hence the qf code :

cdc54b72d5f869d758ba719f5474e8e3ec280fa68b3d7deaf520797d4f52cb4c_9839

where on newer civi versions and because of the change in core mentioned above, the URL would end up like this :

https://test-site-url.co.uk/civicrm/payment/ipn?processor_id=5&cid=62817&conid=36637&mo=contribute&v=testprov&qf=CRMContributeControllerContribution57zmwxhb14owkc4s8kog0sss8owcowwwgogk0g8wscokgkoog4_1941

hence the qf code is now longer, since it now has a prefix (Which is
CRMContributeControllerContribution):

CRMContributeControllerContribution57zmwxhb14owkc4s8kog0sss8owcowwwgogk0g8wscokgkoog4_1941

and thus the resulting URL character count is now longer from what Sagepay can handle (Sagepay does not say anything about this limit in its documentation, but according to my testing it seems to be 202 characters is the max limit), and thus Sagepay throw "3037 : The NotificationURL is too long" error that we can see in the logs :

Exception: A fatal error was triggered: The following errors occurred when submitting payment to Sage Pay:<br />3037 : The NotificationURL is too long.<br />Please contact the site administrator. in CRM_Core_Error::fatal() (line 319 of /var/www/test-site-url.co.uk/httpdocs/sites/all/modules/contrib/civicrm/CRM/Core/Error.php).

My fix here removes the qf code from the list of parameters that are
going to be to used to generate the IPN URL, because to my understanding
the qf code is only needed if the payment processor is redirecting to a
"contribution page" which is not the case usally when using webforms and
instead we will be redirected to the webform submission page, So sending
the qf code from webform context is not really useful.

And even if it happen that the payment processor will redirect to a
contribution page from a webform context, the qf code generated by this module is fake one that is not
linked to any actual contribution page (actual qf codes that are linked
to actual forms should have records inside civicrm_cache table which is not the case here), and thus (and at least to my
undertanding) it really won't matter if the qf code is there or not,
since it is fake.

Anyway, the reason that I am not submitting this change as PR to the main webform_civicrm rep
is because I am not sure if I am missing anything here. So will do that
once we confirm that this fix does not have any side effects that I am
not aware of at the moment.